Add files via upload
The provided Streamlit application is designed to extract features from uploaded .exe files and perform inference to classify them as either benign or malware. It utilizes libraries such as streamlit, boto3, pefile, and tempfile for various functionalities. The extract_features function extracts features from the uploaded .exe file, including n-gram features, imports, section names, and the number of sections. These features are then sent to a SageMaker endpoint for inference using the invoke_endpoint function. The result of the inference is displayed to the user, indicating whether the uploaded file is classified as benign or malware.